requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
openai
pydantic
python-dotenv
requests
sqlalchemy
pymysql
